Move to custom focusin and focusout events
All blur/focusout and focus/focusin events were replaced with custom WinJS
events by registering via WinJS.Utilities._addEventListener for focusout
and focusin, respectively. These custom events bring IE's native
focusout/focusin behavior to other browsers. Specifically:
  1. They ensure that document.activeElement is set appropriately within
     focusout handlers. In Chrome's native focusout,
     document.activeElement is always BODY which isn't very useful.
  2. They work in Firefox even though it doesn't support focusin/focusout
     natively.
  3. They allow handlers to determine who lost focus and who is gaining
     focus. Native blur/focus only provide half of this information.

Also, we have some code in the unit tests which generates keyboard events.
It generates UIEvents but the keyCode property of a UIEvent is not
settable in Chrome. To fix this, switched to CustomEvent so that we can
override the keyCode property.