Fix #49: Ungrouped grid doesn't work in chrome
Ungrouped grid doesn't work in chrome due to a flexbox bug:
  - issue 345433: Incorrect sizing for nested flexboxes
  - https://code.google.com/p/chromium/issues/detail?id=345433

With nested flexboxes, the inner flexbox's width is proportional to the
number of elements instead of the number of columns.

To work around this, we disable structural nodes for environments that
would repro this bug. Disabling structural nodes prevents us from having
nested flexboxes.