Fix #86: Can click thru NavBar arrows onto commands
When clicking rapidly on a NavBar arrow, you can sometimes click thru the
button and onto the NavBar command behind it, invoking it.

This is because a NavBar command receives a z-index of 1 when it is
focused which causes it to be above the arrow. The fix is to give NavBar
arrows a z-index of 2 so that they're always on top of the commands.