Fix AppBar cross browser unit test "KeyboardingInEmptyAppBar" failure. This is one of the unit tests referenced in #117   that is failing cross browser. When an AppBar is opened, the AppBar moves focus to the first AppBar command in DOM Order, however if the AppBar is empty, it focuses itself.
According to the HTML5 spec, DIV's with no tabIndex set should not be focusable through tabbing, pointerDown, or the element.focus() API. However, in IE a DIV with no tabIndex is able to be focused by pointerDown or the element.focus() API.
Other modern browsers seem to get the behavior in the spec correct, so in those browsers the AppBar is not able to focus itself because since no tabIndex is set, and this unit test fails.

The fix is to check if the user has specified a tabIndex on the element already during the AppBar constructor. If they have not, we explicitly set a tabIndex of -1, which means that an element is not focusable through tabbing, but is focusable through pointerDown and the element.focus() API.