Firefox applies default styles to Firefox specific Pseudo-elements in many intrinsic input elements.
This causes our AppBarCommand buttons to be too wide in Firefox, resulting in incorrect math and failing unit tests around AppBar scalability scenarios. ala #117 and # 173
The end result is that in Firefox the AppBar could overflow full size commands into a second row, before knowing to apply the .win-reduced class, when the window was resized.

The fix is to remove the Firefox specific pseudo element styles from AppBarCommands and to reapply the box-sizing css rule with the -mox prefix.

This closes #173