Prevent chevron from wrapping in Firefox
When putting 2 elements in a button styled with display:flex flex:row,
Firefox renders the elements in the same column rather than the same row.
This breaks the visuals for the hub's interactive headers (the chevron
appears under the title rather than next to it).

This issue is worked around by adding an extra div inside of the button.
The button's flex styles and children are moved into this div.