Fix #136: NavBar: Fix resize tests
Tests were failing in browsers that don't support mselementresize.

Developers who resize a NavBar element should call forceLayout on it so
the resize is detected in browsers which do not support mselementresize.

NavBar also resizes when the window is resized in all browsers.

So the resize tests were updated as follows:
  - If the browser supports mselementresize, we do nothing other than
    resize the element. Otherwise, we fire a window resize event.
  - Each resize test was duplicated to call forceLayout after changing the
    element's size. These tests run in all browsers. They test the scenario
    that the developer wants to change the size of the NavBar without
    changing the size of the window.