Fix #136: NavBar: Fix high contrast and URL tests
The high contrast tests were failing in some browsers because not all
browsers support msHighContrastAdjust. This was fixed by only asserting
against this attribute in browsers that support it.

Assertions against CSS URLs were failing in some browsers because the
browsers normalize the URLs differently. This was fixed by doing the
assertions against normalized URLs.