Fix #127: Get Tooltip tests running in all browsers
Most of the problems stemmed from WWA/IE implementing pointer events while
other browsers implement touch events. The only touch events that seem to
be implemented consistently are:
  - touchstart
  - touchend
  - touchmove

Tooltip relies on pointerover and pointerout for:
  - Showing a tooltip when the user presses and holds on a tooltip element
  - Showing a tooltip when the user hovers their finger onto a tooltip
    element
  - Hiding a tooltip when the user hovers their finger off of a tooltip
    element

There aren't any touch event equivalents for pointerover or pointerout. So
the first scenario was achieved in other browsers by also having
pointerdown (touchstart) show a tooltip. There isn't a good way to
implement the other 2 scenarios without pointerover/out so the other 2
scenarios do not work in other browsers. We are okay with this.

The unit tests which fired pointerover/out had to be modified to fire
alternative events in browsers that don't have pointer events.

--

When a tooltip is triggered, the control notes which input method
triggered the tooltip (_showTrigger). Before this change, when multiple
input methods triggered a tooltip, the control remembered the last method.
This was changed slightly so that it remembers the first method instead.

--

Our translation table used touchleave as the touch equivalent of
pointerout. Because touchleave has different semantics than pointerout and
nobody is currently relying on pointerout mapping to touchleave,
pointerout no longer maps to touchleave. When somebody needs pointerout,
we can think more about the appropriate touch equivalent.

--

Most of the other changes were around adjusting the unit test assertions
to account for slight differences in the return values of the CSS APIs of
the browsers.