ui.js in Microsoft.Phone.WinJS.2.1 should not include two versions of the appbar
While debugging an issue generating metadata for the Phone framework package, I ran into a failure caused by the inclusion of two versions of the Appbar in the Phone framework package.
This is caused by the inclusion of NavBar in https://github.com/winjs/winjs/blob/master/src/js/ui-phone.js (which should not be part of the phone framework package). NavBar depends on AppBar (desktop. This is why the phone + desktop version of AppBar get included.