Fix tooltip tests in Safari. #127
The touchEventTranslator makes properties on the touch eventObject
writable so that they can be overriden. This didn't work in Safari because
in Safari, the properties are not configurable. Switched to a different
technique for overriding eventObject properties which works in Safari and
all other browsers.

Also enabled the tooltip tests in all browsers on saucelabs.

A line of code that positioned the tooltip within _showTooltip was
readded. This line was mistakenly removed in an earlier commit
(aaaf1eb8eaeac85bdea8ccd103b6bc47c2925ae8).

The _handledTouch expando on the eventObject isn't used anywhere so it was
removed.