Make TooltipHidingTests more robust. #127
The TooltipHidingTests are currently flaky. The cause seems to be that the
timeframe in which the tooltip is expected to dismiss is too small. The
tooltip is behaving correctly as long as it dismisses faster than
DEFAULT_MESSAGE_DURATION so I extended the expected timeframe while
keeping it under that threshold.