Don't dynamically add stylesheets in unit tests
In the Rating, Toggle, and Tooltip tests, we dynamically add and remove
stylesheets. Not only is this not needed in our current qunit set up but
it also makes the tests fail intermittently. It seems that the code that
waits for a CSS file to be loaded is not robust. This change fixes some of
the failures we're seeing in the tooltip tests (#127).