Re-enable testDeleteDoesNotLoseFocusRectangle #342
VirtualizedViewTests.testDeleteDoesNotLoseFocusRectangle was disabled in
commit 022ff8c5d4edb6572b87045a351367e533c69e25 because it was failing. I
could only get it to fail when the browser wasn't focused. I re-enabled the
test and added extra logging to make it easy to tell whether or not the
browser has focus if the test fails again.

The reason it fails is that programmatically moving focus via setActive
doesn't cause focus events to fire when the browser doesn't have focus
(IE 11).