Enable some ListView tests outside of IE11 #134
Currently, ListView tests only run in IE11. This change creates a new test
module called ListViewBrowser which contains a subset of the ListView
tests (467 of 809). This module is set up to run in Saucelabs in all
environments except IE11.

When we get all of the ListView tests running under a reasonable time
period in the non-IE11 browsers, we can delete this new test module and
set the ListView module to run in all environments on Saucelabs.

The way ListViewBrowser works is it imports all test files from the
ListView module and then disables a hardcoded subset by monkey patching
LiveUnit.registerClass.