Fix CSS grid feature detection in tests #416
Cell spanning tests were running in Safari despite it not supporting CSS
grid. The problem is that Safari CSSStyleDeclarations have a property for
grid-column (specifically, -webkit-grid-column) despite the CSS engine
ignoring such attributes.

This was fixed by switching to a platform specific check for CSS grid
(checking for the existence of -ms-grid-row). This means that ListView
will only use CSS grid in IE. This is probably what we want because the
ListView's use of CSS grid is based on IE 10/11's implementation. If other
browsers were to implement CSS grid, ListView would have be updated anyway
to make use of the latest CSS grid standard.