ListView: selection background disappears but selection checkmark doesn't, Fixes #356
When you configure the ListView as:
tabBehavior: toggleSelect
itemsReorderable: true
swipeBehavior: false

You should be able to swipe to re-order, but swiping should not change selection (or the visual).

The bug was that dispatchSwipeBehavior was calling _animateSelectionChange, which toggles the win-selected class. Therefore, the item was still selected (per listview.winControl.selection), but visually it did not look good since the win-selected class was removed.