Fix keyboarding with tab key in NavBar outside IE
NavBar was using the IE only event beforeactivate to control which element
received focus when the user hit the tab key. This usage was replaced with
an instance of TabContainer.

The behavior of tabbing into a NavBarCommand was changed slightly.
Previously, when you tabbed in, the button or the split button received
focus depending on which one was last active. Now when tabbing in, the
button always receives focus (and the split button never does). This was
done to avoid the complications of introducing a TabContainer into the
NavBarCommand. As a result, the split button now has a tabindex of -1
instead of 0 and all references to _splitButtonActive were removed.

Closes #479.