ListView panning updates for iOS, Fixes #205
The root cause of the panning performance issues we see in iOS is that we don't get scroll events/or an accurate scroll values while there is inertia, so our virtualization code does not know what items to realize.

We evaluated different options:
-Do nothing
-Remove inertia (by stop setting -webkit-overflow-scroll: touch)
-Fake scrolling with translate and animations. We evaluated two libraries (1) iScroll: winjs-ebang.azurewebsites.net/iscroll/listview.html 2) Scrollability: http://winjs-ebang.azurewebsites.net/scrollability/listview.html, and it look "good/ok" on an iPad Air, but not so well on an iPhone 4s (which is still supported for iOS 8)
-Use larger prefetch buffer: a) no virtualization b) larger front buffer

We decided to go with the last option. This change introduces two private flags that allow us to configure the number of pages ahead/behind the scrolling direction that we want to prefetch. It is currently enabled only in iOS, and the default values are 2 behind, 6 ahead of the panning direction.

You can play with the values of these flags here: http://winjs-ebang.azurewebsites.net/lvlistlayout.html

Before:
http://winjs-ebang.azurewebsites.net/lvlistlayout-prehover.html

After:
http://winjs-ebang.azurewebsites.net/lvlistlayout.html

Demo: http://1drv.ms/1pD0F6T

Note: iOS still allows you to scroll huge amount of pixels (thousands), in those cases, we will still see backdrops (the only option would be huge prefetch buffers which would have memory implications).  We should call out that even in the iOS AppStore, you are able to see blankness on fast-scrolls: http://1drv.ms/1pD0Mzl