Fix #248: Make Hub headers focusable in Safari
Hub headers are designed to receive focus thru their button elements.
However, Safari has special focus behavior for buttons such that they can
only be focused programmatically (so clicking on a button doesn't give it
focus).

To allow users to focus headers in Safari by clicking on them, a div
within the header was made focusable by giving it a tabindex of -1. Also,
this div was given an outline of none to hide Safari's default blue focus
outline. This is similar to the technique used by AppBar commands.