Focus cache
Focus cache (_focusOrder) sorts the light dismissables such that the
more recently a dismissable had focus, the closer it is to the front.
This is used when calculating the next topLevel light dismissable. It is
the light dismissable that had focus most recently and is above the
click eater.

This fixes the following scenario:
1. Open just the top AppBar and make it sticky
2. Use the "Show" button on the button to open a Flyout
3. Dismiss the flyout

As expected, the "Show" button now regains focus. Previously, the top
AppBar would have regained focus (Because it is at the top of the stack.
However, it is not the only dismissable above the click eater.).