Fix #583: Unit test failure: Security_markSupported
There was a bug in the test. Chrome lowercases attributes set with
setAttribute while IE does not. The test was relying on setAttribute not
changing the casing of the attribute.

The fix was to change the test to use an all lowercase attribute. Other
than that, the attribute's name doesn't matter.