Fix #683: SplitView: Position relative descendants
Currently, if you put an element with position: relative/absolute inside
of the SplitView's content, this element will end up being rendered on top
of the overlayed pane when it opens. For example, you can see this by
putting a SearchBox inside of the SplitView's content.

To avoid this, the SplitView's pane and content are each given a z-index
so that:
  1. They are each a stacking context so their descendants can't escape
     them.
  2. The pane element is always rendered on top of the content element.