Fix ContentDialog dispose
Disposing a hidden ContentDialog would lead to an exception which would
cause the app to terminate.

Here's why. When disposing a hidden dialog, the dialog would put the show
promise into an error state. Because the dialog wasn't showing, nobody
outside of the dialog would have a reference to the show promise so nobody
could handle the error promise. Thus, the promise error would remain
unhandled and lead to app termination.

The fix is to only create the show promise when somebody asks for it by
calling the show() function.

To catch this kind of mistake in the tests, added initUnhandledErrors and
validateUnhandledErrors to all of the ContentDialog tests.

Also, removed overflow:hidden from the dialog's title so that low hanging
letters don't get clipped (clipping used to be needed when we wanted to
limit the number of lines of text in the title).