Update NavBar styles to match new redlines
NavBar's layout logic had to be updated. In horizontal layout, each
command should be separated by 4px of horizontal space. However, the first
and last command in each row shouldn't have any left margin or right
margin, respectively. To acheive this, the NavBar's layout logic has to
special case the first and last command of each row to remove its
appropriate margin.

color-definitions.less was updated. @colorMid and @colorMediumMid are
currently the same and that's okay. We want to keep them both in case we
want to change them to different values in the future.

Also updated some NavBar tests:
- VerticalKeyboarding: Had to be updated due to changes in the heights of
  the viewport and commands
- DataEdits, ArrowKeysPressed: Made these tests faster (they were quite
  slow in Firefox)