Closes #726 https://github.com/winjs/winjs/issues/726
This is a temporary fix that removes the toolbar command's dependancy for click handling off of its super class. This will likely get reverted with the next checkin for cascading menu support as that code has been stabilized and the super class' click handler will once again be hospitable for toolbar commands.