SplitView: Fix #801 percentage heights in content
An element with a percentage height (e.g. 100%) which was placed inside of
the SplitView's content element did not receive the correct height in
Chrome. This change works around the Chrome bug that caused this behavior.

Also, updated SplitView's usage of .flex to follow a pattern which does a
better job of conveying intent:
  - Elements which should size to content were changed to .flex(none)
  - Elements which should flex were given an explicit basis of 0%. This is
    necessary until we change the flexbox mixin's default flex basis to 0%
    as described in #802.