Closes #798 https://github.com/winjs/winjs/issues/798
The solution is to change the implementation of AppBar's "getCommandByID" API to look through its _layout's "commandsInOrder" property for the id in question, rather than doing a querySelector on the AppBar element for the id. This is because menu layout, removes hidden commands from the DOM, so an element querySelector does not suffice.