Closes #840 https://github.com/winjs/winjs/issues/840
We were only setting up the flyout for custom content when shownDisplayMode was "reduced". The fix was to set it up regardless.
Additionally, we updated the ToolBar Menu command to update the "selected command" whenever a command was invoked by any means, rather than only when it was clicked, to fix the "null of undefined property" exception we were seeing.