Force light dismissables to have a tabIndex
Every light dismissable should have a tabIndex on its root element. Here's
why. When clicking on a non-focusable element, the browser will move focus
to the first focusable ancestor. If none of the ancestors has focus, then
the browser will move focus to the body. Because focus has moved outside
of the light dismissable, it will dismiss. To prevent this undesirable
behavior, every light dismissable should have a tabIndex on its root
element.