Closes #882 https://github.com/winjs/winjs/issues/882
This is caused by the reentrancy lock of the Cascade Manager. Our desire is to finish clearing all references the current cascade holds to the hiding Flyout and any of its subFlyouts before introducing a showing Flyout to the cascade.

Today we are correctly queueing the showing flyout in hopes that it will try and show again when the reentrancy lock is cleared, but we never schedule an operation to drain the queue upon clearing the lock.