Fix #146: Replace data during lazy container creation
A "Container missing after updateContainers" exception was thrown in the
following app scenario:
  - ListView is initialized with local data and a fetch is made to get the
    latest data from the server.
  - While the ListView is lazily creating containers, the server responds.
    The app replaces the local data in the ListView's data source with the
    data from the server. Specifically, this can be done by setting the
    Binding.List's length to 0 and pushing on the data from the server.

The problem was that the container creation set a cap at the current
number of containers + the change in the data source count. When doing a
direct replacement of the Binding.List's data, the count doesn't change so
updateContainers doesn't create any new containers. However, it receives a
number of notifications for items it hasn't created containers for yet
(because lazy container creation was still in progress). These items
require containers for animations. Because these items don't get
containers, the exception gets thrown.

The fix is to set the container creation cap at the current number of
containers + the number of inserted notifications we received (when lazy
container creation is in progress). This will ensure that updateContainers
creates enough containers for the animations.