ListView: Cross-browser method for hiding scrollbar
During ListView's entrance/exit animation, we don't want to animate the
scrollbar. We solve this differently in different browsers:
  - IE: Animate the win-viewport and hide scrollbar with
    -ms-overflow-style:none (so that the scrollbar isn't animated).
  - Other browsers: Instead of hiding the scrollbar, we animate
    win-surface so that the scrollbar is visible but doesn't move.

The technique of animating win-surface causes a bug where items may be
positioned incorrectly because during the entrance/exit animation,
win-surface becomes the offsetParent of the win-containers. The ListView
is written under the assumption that win-viewport will always be the
offsetParent of the win-containers.

Specifically, this can be triggered by editing the ListView while it's
still playing its entrance animation so that the entrance animation and
editing animations are played at the same time. The repro from
https://github.com/winjs/winjs/issues/146#issuecomment-41634154 triggers
this.

To fix this, all browsers now:
  - Play the entrance/exit animation on win-viewport
  - Hide the scrollbar during the animation by setting overflow:hidden on
    win-viewport.

This change doesn't cause a measurable change in performance during the
entrance/exit animations.