Fix #891: ListView items sometimes jump during animations
At the start and end of editing animations, itemsblocks jump by the amount
of the win-container's top margin. The bug is due to the win-container's
top margin collapsing thru the win-itemcontainer during animations.
This margin collapsing only occurs during animations (due to
win-itemscontainer switching between overflow:hidden and visible). Whether
or not the margin collapsing is happening affects the size of the
win-itemscontainer which is what causes the jumping.

The fix is to prevent the margin collapsing from happening during
animations. The win-itemscontainer-padder prevents the margin collapsing
as long as it isn't display:none. Instead of styling display:none, we now
style height and width to 0. Consequently, win-itemscontainer-padder
prevents margin collapsing without taking up any space.

Most of the change was to update the tests to catch this bug. The
animation tests would have caught this bug if they verified the animation
state of a greater number of items. The changes included:
  - Verifying the size and position of all visible items (instead of just
    the deleted item)
  - Extending the height and width of the ListView so that more items fit
    on the screen.
  - Decreasing the number of items per group so that more groups fit on the
    screen.
  - Instead of making the deleted item the first or last element on the
    screen, put it closer to the middle of the screen so that there will
    be some items above it and some items below it that need to be
    animated.
  - Styling win-container's margin-top instead of margin-bottom (this is
    what we recommend to apps to avoid margin collapsing in vertical
    ListLayout).
  - Updated styling of win-itemscontainer's margins to reflect the fact
    that the defaults are never -200px anymore.

Deleted some dead test code.