Fix #976 Build output doesn't have BOMs
The problem was that the line-ending task ran after the add-bom task and
happened to remove the BOM. Swapping the order of these tasks causes the
build output to have both a BOM and the correct line endings.

It would be nice to have a build task that runs during "release" that
fails the build if the output is missing a BOM or has the wrong line
endings.