Enable WinJS.js output file to work with webpack
The WinJS.js output file didn't work with webpack. The problem was that
our technique for getting ahold of the global object didn't work:

  (function (global) {
    // In here, we wanted global to be the global object.
  })(this);

This techinque failed because webpack executes a module's code within a
function where "this" is not the global object.

We now use this technique:

  var globalObject =
    typeof window !== 'undefined' ? window :
    typeof self !== 'undefined' ? self :
    typeof global !== 'undefined' ? global :
    {};

which gets ahold of the appropriate global object whether its executed on
the browser UI thread, a browser web worker, or node even when used with
webpack.