Previous versions of WinJS have relied on logic contained in the private _Overlay class to ensure that they remain visible by handling the IHM "showing" and "hiding" events and adjusting their offsets accordingly.
Now that the new AppBar Class no longer inherits from Overlay, the _keyboardInfo helper object has been extracted from the _Overlay class and made into a new private stand alone module called _KeyboardInfo. Both the _Overlay class and the new AppBar class consume it.

This change provides the AppBar with the means to not stay covered by the IHM whenever the IHM shows and covers the bottom portion of the App window.