Fix Menu Command with Flyout
When using a menu command with a flyout in a toolbar (as overflow for content or otherwise) the flyout is shown on hover. 
Invoking the MenuCommand._activateFlyoutCommand again when the flyout is shown causes an exception. 
This change checks if the flyout associated with the MenuCommand is hidden before calling that function.