Makes the menu command aware of flyout positioning
This makes the `MenuCommand` check it's position relative to the edge of the document before showing a flyout on the right. If the menu is all the way on the right side of the screen, showing the flyout to right is not a good option. 
This change adds a check to see if more than 100px are available to the right of the `MenuCommand`