Fix #1081: ListView layout doesn't clean up envInfo
After disposing a ListView, when attempting to reuse its layout with a new
ListView, the new ListView doesn't render properly. Repros in browsers that
don't support CSS grid.

Here's a description of the problem and fix. Layouts can be reused. They
have 2 lifecycle methods: initialize and uninitialize. The problem here is
that the layout isn't properly cleaning up some state during uninitialize.
Specifically:
  - The layout should clear its envInfo.
  - The layout should remove the win-nocssgrid class from the surface

The former point is causing the symptom. The layout adds the win-nocssgrid
class to the ListView when it calculates envInfo but in the new ListView,
it already has envInfo so it never adds win-nocssgrid to it causing it to
render incorrectly.