Fix #1133: Fix require('winjs') in CommonJS env
This:

  var WinJS = require('winjs');

now works when using WinJS in a CommonJS environment (e.g. browserify).

All of the changes were in requirejs.js which generates the JavaScript at
the head and tail of base.js and ui.js.

To fix this:
  - ui had to declare its dependency on base.js CommonJS-style. ui.js now
    starts with the following code which declares its dependency on base for
    all module styles: AMD, CommonJS, and global/none:

    if (typeof define === 'function' && define.amd) {
      // amd
      define(["./base"], factory);
    } else {
      if (typeof module !== 'undefined') {
        // CommonJS
	factory(require("./base"));
      } else {
        // No module system
	factory(globalObject.WinJS);
      }
    }

  - ui and base had to publish their exports CommonJS-style. ui.js now
    ends with the following code which publishes its exports for all module
    styles: AMD, CommonJS, and global:

    // WinJS always publishes itself to global
    globalObject.WinJS = _WinJS;
    if (typeof module !== 'undefined') {
      // This is a CommonJS context so publish to exports
      module.exports = _WinJS;
    }
    return globalObject.WinJS;
  - js/ui.js is listed as "main" in package.json. Because ui.js depends on
    base.js, this will pull in all of WinJS.

I tested this change with browserify and webpack.

Note that WinJS bundles created with browserify currently lead to an
exception in the browser due to stray BOMs in the file. This is due to
substack/node-browserify#1263. It can be worked around by running
browserify with this flag: --no-detect-globals