Fix #1202: Accent color sometimes not applied
Some times none of the accent color rules work because they are generated
with an accent color value of 'undefined'. This is due to a race condition
in the accent color code: sometimes the accent color rules are generated
before the accent color has been retrieved.

The core of the fix is for the rule generator (scheduleWriteRules) to wait
for the accent color to be retrieved before writing the rules.

There's more red/yellow in the diff than needed due to some code being
moved around so here's a summary of the changes:
  - Put code for retrieving accent color into a function called
    'initializeColors'.
  - Put event handler (handleColorsChanged) for responding to a change in
    the accent color inside of the 'initializeColors' function. The reason
    being this makes it very clear that initializeColors is the only one
    that uses it and there's no way it can be called before the accent
    color has been initialized.
  - Create a promise (initializeColorsPromise) that completes when
    'initializeColors' is finished.
  - Have 'scheduleWriteRules' wait for 'initializeColorsPromise' to
    complete.
  - Switch the setImmediate in 'scheduleWriteRules' to a Promise.timeout()
    so it's easier to compose with the other promise
    (initializeColorsPromise).