Fix #1202: Accent color sometimes not applied
Some times none of the accent color rules work because they are generated
with an accent color value of 'undefined'. This is due to a race condition
in the accent color code: sometimes the accent color rules are generated
before the accent color has been retrieved.

The fix makes accent color retrieval synchronous. In addition to fixing
this race condition, it also fixes an issue where the accent color could
flash in. If it took a while to reach DOMContentLoaded, the user would
initially see the page rendered without accent color and then after
DOMContentLoaded, they'd see the page rendered with accent color. This is
doesn't happen anymore beacuse color retrieval no longer waits on
DOMContentLoaded.