Make accent color retrieval synchronous
When it was async, the page would initially load without accent color and
then all of a suddent accent color would be applied when DOMContentLoaded
fired.

We made accent color retrieval synchronous to fix this. The catch is we
have to put an element in document.head instead of document.body (which is
only available after DOMContentLoaded) and query its style to see if WinJS
is in the dark theme or light theme.