Introduce _ElementUtilities._getComputedStyle helper
Firefox's implementation of getComputedStyle returns null when called
within an iframe that is display:none. This is a Firefox bug. Most of the
WinJS code expects getComputedStyle to return an object whose keys map to
strings.

To work around this Firfeox bug, introduced a helper
(_ElementUtilities._getComputedStyle). It always returns an object whose
keys map to strings.

Also fixed #1312 because it was breaking the unit test that this change
was adding. Essentially, MessageEvent.source can be undefined in Chrome
(when the iframe has already been removed from the DOM) so we have an
extra if statement to protect against this case.